identificator/requirements.txt
grpcio==1.48.2
grpcio-tools==1.48.2
protobuf==3.19.6
pymongo==4.1.1
six==1.16.0
